Make SSH command connection functions public
